+++ src/SimpleStocker.Api/Services/ProductService.cs
@@ -11,9 +11,11 @@ namespace SimpleStocker.Api.Services
     public class ProductService : IProductService
     {
         private readonly IProductRepository _repository;
-        public ProductService(IProductRepository repository)
+        private readonly ISaleItemRepository _saleItemRepository;
+        public ProductService(IProductRepository repository, ISaleItemRepository saleItemRepository)
         {
             _repository = repository;
+            _saleItemRepository = saleItemRepository;
         }
 
         public async Task<ApiResponse<ProductViewModel>> CreateAsync(ProductViewModel entity)
@@ -28,7 +30,7 @@ namespace SimpleStocker.Api.Services
                 var res = await _repository.CreateAsync(mapperModel);
                 if (res == null)
                     return new ApiResponse<ProductViewModel>("Server", "Erro ao tentar criar registro!");
-                return new ApiResponse<ProductViewModel>();
+                return new ApiResponse<ProductViewModel>(true, "", [], Mapper.Map<ProductViewModel>(res), 200);
             }
             catch (Exception ex)
             {
@@ -43,6 +45,11 @@ namespace SimpleStocker.Api.Services
                 var foundEntity = await _repository.GetOneAsync(id);
                 if (foundEntity == null)
                     return new ApiResponse<ProductViewModel>("Id", "Id não encontrado!");
+
+
+                var sales = await _saleItemRepository.GetAllSaleItemsByProductId(id);
+                if(sales.Count>0)
+                    return new ApiResponse<ProductViewModel>("SaleItem", "Existem vendas vinculadas com esse produto!");
 
                 var deleteItem = await _repository.DeleteAsync(foundEntity);
                 if (deleteItem)
@@ -96,7 +103,7 @@ namespace SimpleStocker.Api.Services
         public async Task<ApiResponse<ProductViewModel>> UpdateAsync(ProductViewModel entity)
         {
             var originalEntity = await _repository.GetOneAsync(entity.Id);
-            if (originalEntity == null)
+                if (originalEntity == null)
                 return new ApiResponse<ProductViewModel>("Id", "Item não encontrado");
 
             var validation = new ProductValidator(true).Validate(entity);
@@ -110,7 +117,7 @@ namespace SimpleStocker.Api.Services
                 var res = await _repository.UpdateAsync(mapperModel);
                 if (res == null)
                     return new ApiResponse<ProductViewModel>("Server", "Erro ao tentar criar registro!");
-                return new ApiResponse<ProductViewModel>();
+                return new ApiResponse<ProductViewModel>(true, "", [], Mapper.Map<ProductViewModel>(res), 200);
             }
             catch (Exception ex)
             {
